*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@4074 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/T.c
@@ -265,7 +265,7 @@ long T_atol( const char *txt )
 /*---------------------------------------------------------------------*/
 /*---------------------------------------------------------------------*/
 
-long T_atoi( const char *txt )
+int T_atoi( const char *txt )
 {
 	long ret;
 
@@ -277,7 +277,7 @@ long T_atoi( const char *txt )
 		THROW( EXCEPTION );
 	}
 
-	return ret;
+	return ( int ) ret;
 }
